refactor: replace live_docs_cache with live_docs_fetch
The server's caching was designed for a long-running process handling
concurrent requests. As a one-shot deployment script, all we need is
in-process memoization to avoid redundant HTTP calls within a single
run. Replace the threading/warmup machinery with functools.lru_cache
on the four fetch functions, and rename the file accordingly.
